\name{rtk-package}
\alias{rtk-package}
\docType{package}
\title{
\packageTitle{rtk}
}
\description{
\packageDescription{rtk}
}
\details{

The DESCRIPTION file:
\packageDESCRIPTION{rtk}
\packageIndices{rtk}
This package might be used to rarefy data and compute diversity measures. Rarefied tables can be returned to R and be further processed.
}
\author{
\packageAuthor{rtk}

Maintainer: \packageMaintainer{rtk}
}
\references{
~~ Literature or other references for background information ~~
}

\seealso{
\code{\link{rtk}}, \code{\link{plot.rtk}}, \code{\link{collectors.curve}
}
}
